Adds QMSaveAnnotation for classical outcome saving.
Adds a new annotation, `QMSaveAnnotation`, to signal the backend/QPU compiler to materialize stream-processing saves for classical entities within a BoxOp.

This annotation enables offline decoding without constraining real-time reuse of classical bits.

Also, it introduces the `QMSaveOpenQASM3Serializer` for serializing the annotation to OpenQASM3.

The `QMBackend` is updated to include the name of the backend in its target description.

The `Direction` enum now supports a `BOTH` option, indicating bidirectional data flow between DGX_Q and OPX.
This also updates the parameter and parameter table to handle bidirectional streams correctly.