Removed FluxChannel to the benefit of more ControlChannels for visual
FluxChannel previously introduced is not plotted by default by IBM tools, FluxChannels are now cast as other ControlChannels (extra to the ones for the potential couplers). The user should just access them from the qubit perspective anyway so that should be fine.

Furthermore, added an extra property on the InstructionProperties to access easily the Quam macro object itself on top of just the callable. Can be useful to fetch some properties of it.