Enables simulation for estimator and sampler jobs
Adds functionality to run estimator and sampler jobs in simulation mode
using the Quantum Machines Manager (QMM) simulator. It checks if the
backend is a QMM instance and the 'simulate' option is provided in the
job metadata. If so, it uses the simulator to run the job.

Also adds properties to the `QmSaasProvider` class to expose the client,
version, and instance objects. It adds the `close_all` method to close
all QmSaas instances and QuantumMachinesManager instances, as well as a
`spawn` method to spawn a new QmSaas instance.